+++ hyperfit/src/main/java/org/hyperfit/net/ResponseInterceptor.java
@@ -1,0 +1,10 @@
+package org.hyperfit.net;
+
+/**
+ * Intercept response.
+ */
+public interface ResponseInterceptor {
+
+    public void intercept(Response response);
+
+}